refactor: outsourced createEmailToken
